requirements.txt
Flask~=2.1.2
opencv-python~=4.6.0.66
numpy~=1.23.1
Werkzeug~=2.1.2
keras~=2.9.0
tensorflow~=2.9.1
pre-commit~=2.20.0
pytest~=7.1.2
Pillow~=9.0.1
modAL~=0.4.1
scikeras~=0.9.0
imgaug~=0.4.0
matplotlib~=3.5.2
tf-explain